\name{interact}
\alias{interact}
\title{Computes interactions between modules.}
\usage{
interact(dat, sp_col, bioregion_col, score_col, plot = FALSE,
output_format = "matrix")
}
\arguments{
\item{dat}{a .tp object generated by OSLOM algorithm. Import in R with
readLines() function. This object can be generated with run_oslom() function.}

\item{sp_col}{a character string corresponding to the column
    with species.}

\item{site_col}{a character string corresponding to the column
    with sites.}

\item{bioregion_col}{a character string corresponding to the column
    with bioregions.}
}
\description{
See Guimera et Amaral, Nature, 2005.
}
\details{}
\section{Note}{}

\examples{
ex <- data.frame(sites = c(rep("A", 2), rep("B", 3), rep("C", 2)),
species = c("a", "b", "a", "c", "d", "b", "d"),
count = c(10, 100, 1, 20, 50, 10, 20),
bioregion = c(rep("region1", 5), rep("region2", 2)))

z_scores <- zscore(ex, sp_col = "species", site_col = "sites",
                   bioregion_col = "bioregion")
z_scores
}
