\name{oslom_output}
\alias{oslom_output}
\title{Convert the .tp output of OSLOM algorithm into a data.frame,
with each site having its bioregion.}
\usage{
oslom_output(dat, contingency_mat)
}
\arguments{
\item{dat}{a .tp object generated by OSLOM algorithm. Import in R with
readLines() function. This object can be generated with run_oslom() function.}

\item{contingency_mat}{a contingency table with sites as rows
and species as columns}

}
\description{
Convert the .tp output of OSLOM algorithm into a data.frame,
with each site having its bioregion.
WARNING: the order of the rownames of the contingency table
has to be identical before running OSLOM and when running this function!
}
\details{}
\section{Note}{}

\examples{
}
