\name{zscore_dplyr}
\alias{zscore_dplyr}
\title{Compute zscores for species from a given delineation of sites.}
\usage{
zscore_dplyr(dat, sp_col, site_col, bioregion_col)
}
\arguments{
\item{dat}{a .tp object generated by OSLOM algorithm. Import in R with
readLines() function. This object can be generated with run_oslom() function.}

\item{sp_col}{a character string corresponding to the column
    with species.}

\item{site_col}{a character string corresponding to the column
    with sites.}

\item{bioregion_col}{a character string corresponding to the column
    with bioregions.}
}
\description{
See Guimera et Amaral, Nature, 2005.
}
\details{}
\section{Note}{}

\examples{
ex <- data.frame(sites = c(rep("A", 2), rep("B", 3), rep("C", 2)),
species = c("a", "b", "a", "c", "d", "b", "d"),
count = c(10, 100, 1, 20, 50, 10, 20),
bioregion = c(rep("region1", 5), rep("region2", 2)))

z_scores <- zscore_dplyr(ex, sp_col = "species", site_col = "sites",
                   bioregion_col = "bioregion")
z_scores
}
